nixos/malloc: warn about old Scudo options
NixOS 25.11 uses standalone Scudo for the first
time (85b124c). With this change, options are now
snake_case instead of CamelCase.

https://llvm.org/docs/ScudoHardenedAllocator.html#options

There is currently no breaking-change warning or
evaluation warning to alert users of this change.
As a consequence, users may run into scudo's runtime
warnings. This adds a straightforward evaluation-time
warning if an old-style option is detected.